refactor: add id to base entity
