+++ domain/src/main/java/nexters/payout/domain/BaseEntity.java
@@ -2,6 +2,7 @@ package nexters.payout.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
+import nexters.payout.domain.dividend.domain.Dividend;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
@@ -14,6 +15,9 @@ import java.util.UUID;
 @EntityListeners(AuditingEntityListener.class)
 @Getter
 public class BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
 
     @Column(name = "created_at", updatable = false)
     @CreatedDate
@@ -23,4 +27,17 @@ public class BaseEntity {
     @LastModifiedDate
     private Instant lastModifiedAt;
 
+    public BaseEntity(UUID id) {
+        this.id = id;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        return obj instanceof BaseEntity && this.id.equals(((BaseEntity) obj).getId());
+    }
 }
